+++ src/soag/eSOAG.d
@@ -209,9 +209,9 @@ void Error(T)(int ErrorType, T Proc)
 
 void Expand() nothrow @safe
 {
-    long NewLen(long ArrayLen)
+    size_t NewLen(size_t ArrayLen)
     {
-        if (ArrayLen < DIV(int.max, 2))
+        if (ArrayLen < DIV(size_t.max, 2))
             return 2 * ArrayLen + 1;
         else
             assert(0);
